requirements.txt
psycopg2-binary
openai~=1.50.2
python-dotenv~=1.0.0